requirements.txt
anyio==3.5.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
Django==4.0.4
djangorestframework==3.13.1
dnspython==2.2.1
elastic-transport==8.1.2
elasticsearch==8.2.0
email-validator==1.1.3
fastapi==0.75.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
limits==2.6.2
MarkupSafe==2.1.1
orjson==3.6.8
packaging==21.3
passlib==1.7.4
pika==1.2.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
pylogbeat==2.0.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.20.0
python-logstash==0.4.8
python-logstash-async==2.5.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
starlette==0.17.1
typing-extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
zipp==3.8.0
